Parsing of objects is now working for all objects. Made some changes to header files; mostly changed integers to floats.

+++ include/raycast.h
@@ -5,7 +5,7 @@
 struct sphere {
     pixel color;
     v3 pos;
-    u32 rad;
+    float rad;
 };
 
 struct plane {
